refactor: 💡 Split `ca-checkpoint.ts` into separate files
